v5 스키마 및 도메인 규칙에 맞춘 Model/DAO 개편
- 식별자 및 단위 타입을 `int` / `Integer`로 일괄 표준화하여 가독성 및 효율성 향상
- 금액 및 대형 수치 데이터는 오버플로우 방지를 위해 `long` / `Long` 유지
- 데이터베이스 제약조건(`NOT NULL` vs `NULL 허용`)에 맞춰 자바 타입을 원시형과 래퍼 클래스로 철저히 분리 적용
- PreparedStatement의 `setObject()` 및 ResultSet의 `wasNull()` 처리를 통해 NULL 값 데이터 무결성 완벽 보장
- 각 도메인별 DAO 클래스에 일관된 싱글톤 패턴 적용 및 스키마 기반 상세 주석 보강
- 적용 범위: erp.employees, erp.payroll, erp.retirement, erp.settings 패키지 내 모델 및 DAO 전수 갱신

+++ src/main/java/erp/employees/model/EmployeeAppointment.java
@@ -8,8 +8,8 @@ import java.util.Date;
 public class EmployeeAppointment {
 
 	// [DB 관리 항목]
-	private Long employeeAppointmentId; // 인사발령 이력 식별 번호 (PK)
-	private Long employeeId; // 사원 식별 번호 (FK)
+	private int employeeAppointmentId; // 인사발령 이력 식별 번호 (PK)
+	private int employeeId; // 사원 식별 번호 (FK)
 
 	// [인사발령 입력 항목]
 	private String appType; // 발령구분 (승진, 부서이동 등)
@@ -23,19 +23,19 @@ public class EmployeeAppointment {
 	}
 
 	// Getter & Setter
-	public Long getEmployeeAppointmentId() {
+	public int getEmployeeAppointmentId() {
 		return employeeAppointmentId;
 	}
 
-	public void setEmployeeAppointmentId(Long employeeAppointmentId) {
+	public void setEmployeeAppointmentId(int employeeAppointmentId) {
 		this.employeeAppointmentId = employeeAppointmentId;
 	}
 
-	public Long getEmployeeId() {
+	public int getEmployeeId() {
 		return employeeId;
 	}
 
-	public void setEmployeeId(Long employeeId) {
+	public void setEmployeeId(int employeeId) {
 		this.employeeId = employeeId;
 	}
