+++ src/main/java/erp/settings/model/AttendanceGroup.java
@@ -1,22 +1,22 @@
 package erp.settings.model;
 
 /**
- * [기본환경설정] 휴가/근태 설정 - 근태항목 설정 (근태그룹) Model DB 테이블: ATTENDANCE_GROUP
+ * [기본환경설정] 휴가/근태 설정 - 근태그룹 설정 Model DB 테이블: ATTENDANCE_GROUP
  */
 public class AttendanceGroup {
 
-	private Long attendanceGroupId;
+	private int attendanceGroupId; // 근태그룹 식별 번호 (PK)
 	private String groupName;
 
 	public AttendanceGroup() {
 	}
 
 	// Getter & Setter
-	public Long getAttendanceGroupId() {
+	public int getAttendanceGroupId() {
 		return attendanceGroupId;
 	}
 
-	public void setAttendanceGroupId(Long attendanceGroupId) {
+	public void setAttendanceGroupId(int attendanceGroupId) {
 		this.attendanceGroupId = attendanceGroupId;
 	}
